requirements.txt
beautifulsoup4
docx2txt
langchain
langchain_community
langchain_text_splitters
openai
openpyxl
pandas
pytz
Requests
xlrd
faiss-cpu
websocket-client
streamlit
virtualenv
tiktoken
transformers
transformers_stream_generator
optimum
pdfplumber
wikipedia
arxiv
bitsandbytes
accelerate
fastapi
py-cpuinfo
diskcache
requests_toolbelt
playsound
tabulate
chardet
tenacity
cpm_kernels
pydub
keyboard
sounddevice
scipy
gradio_client
neo4j
ChatTTS
soundfile
jax
python-dotenv